+++ tea-coffee-core/index.php
@@ -11,7 +11,7 @@
  *
  * @package Wordpress
  * @subpackage Tea Coffee
- * @version 1.0
+ * @version 1.2.3
  */
 
 get_header(); ?>