+++ tea-coffee-core/sidebar.php
@@ -6,7 +6,7 @@
  *
  * @package Wordpress
  * @subpackage Tea Coffee
- * @version 1.2.2
+ * @version 1.2.3
  */
 
 if ( ! is_active_sidebar( 'sidebar1' ) ) {